chore: remove redundant files, expect watermark data in source
