--- ios/DorisTypesMappers/AVDorisMetadataMapper.swift
@@ -1,22 +1,0 @@
-//
-//  AVDorisMetadataMapper.swift
-//  react-native-video
-//
-//  Created by Yaroslav Lvov on 10.06.2021.
-//
-
-import AVDoris
-
-//class AVDorisMetadataMapper {
-//    func map(metadata: Metadata?) -> DorisUIMetadataConfiguration? {
-//        guard let metadata = metadata else { return nil }
-//        
-//        let dorisTranslations = DorisUIMetadataConfiguration(title: metadata.title,
-//                                                             infoDescription: metadata.description,
-//                                                             type: metadata.type,
-//                                                             thumbnailUrl: metadata.thumbnailUrl,
-//                                                             channelLogoUrl: metadata.channelLogoUrl)
-//        
-//        return dorisTranslations
-//    }
-//}
