--- ios/Helpers/AppIdFetcher.swift
@@ -1,17 +1,0 @@
-//
-//  AppIdFetcher.swift
-//  RNDReactNativeDiceVideo
-//
-//  Created by Yaroslav Lvov on 02.12.2020.
-//
-
-import Foundation
-
-class AppIdFetcher {
-    static let shared = AppIdFetcher()
-    private var appID: Int?
-    
-    func fetchAppId(completion: @escaping (String?) -> Void) {
-        completion(Bundle.main.bundleIdentifier)
-    }
-}
