+++ ios/Video/RCTVideoManager.swift
@@ -8,16 +8,10 @@
 
 import Foundation
 
-#if OldPlayer
-typealias CurrentPlayerView = PlayerView
-#else
-typealias CurrentPlayerView = NewPlayerView
-#endif
-
 @objc(RCTVideoManager)
 class RCTVideoManager: RCTViewManager {
     override func view() -> UIView! {
-        let view =  CurrentPlayerView()
+        let view = NewPlayerView()
         view.jsBridge = bridge
         
         let weirdViewThatCausesFocusIssues = bridge.uiManager.view(forReactTag: NSNumber(integerLiteral: 25))
@@ -29,7 +23,7 @@ class RCTVideoManager: RCTViewManager {
     //MARK: Differs (ios only)
     @objc public func seekToNow(_ node: NSNumber) {
         DispatchQueue.main.async {
-            let component = self.bridge.uiManager.view(forReactTag: node) as? CurrentPlayerView
+            let component = self.bridge.uiManager.view(forReactTag: node) as? NewPlayerView
             component?.seekToNow()
         }
     }
@@ -37,21 +31,21 @@ class RCTVideoManager: RCTViewManager {
     //MARK: Differs (ios only)
     @objc public func seekToTimestamp(_ node: NSNumber, isoDate: String) {
         DispatchQueue.main.async {
-            let component = self.bridge.uiManager.view(forReactTag: node) as? CurrentPlayerView
+            let component = self.bridge.uiManager.view(forReactTag: node) as? NewPlayerView
             component?.seekToTimestamp(isoDate: isoDate)
         }
     }
     
     @objc public func seekToPosition(_ node: NSNumber, position: Double) {
         DispatchQueue.main.async {
-            let component = self.bridge.uiManager.view(forReactTag: node) as? CurrentPlayerView
+            let component = self.bridge.uiManager.view(forReactTag: node) as? NewPlayerView
             component?.seekToPosition(position: position)
         }
     }
     
     @objc public func replaceAdTagParameters(_ node: NSNumber, payload: NSDictionary) {
         DispatchQueue.main.async {
-            let component = self.bridge.uiManager.view(forReactTag: node) as? CurrentPlayerView
+            let component = self.bridge.uiManager.view(forReactTag: node) as? NewPlayerView
             
             var adTagParameters = [String: Any]()
             var startDate: Date?
@@ -77,14 +71,14 @@ class RCTVideoManager: RCTViewManager {
     
     @objc public func seekToResumePosition(_ node: NSNumber, position: Double) {
         DispatchQueue.main.async {
-            let component = self.bridge.uiManager.view(forReactTag: node) as? CurrentPlayerView
+            let component = self.bridge.uiManager.view(forReactTag: node) as? NewPlayerView
             component?.setInitialSeek(position: position)
         }
     }
     
     @objc public func limitSeekableRange(_ node: NSNumber, payload: NSDictionary) {
         DispatchQueue.main.async {
-            let component = self.bridge.uiManager.view(forReactTag: node) as? CurrentPlayerView
+            let component = self.bridge.uiManager.view(forReactTag: node) as? NewPlayerView
             if let limitedSeekbleRange = try? Source.LimitedSeekableRange(dict: payload) {
                 component?.setupLimitedSeekableRange(with: limitedSeekbleRange)
             }
