+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "react-native-video",
-    "version": "7.1.5",
+    "version": "7.2.0",
     "dorisAndroidVersion": "3.6.12",
     "description": "A <Video /> element for react-native",
     "main": "Video.tsx",
@@ -20,7 +20,7 @@
         "eslint-plugin-react": "3.16.1"
     },
     "dependencies": {
-        "@dicetechnology/react-native-dice-video": "git+ssh://git@github.com:DiceTechnology/react-native-dice-video.git#6.42.4",
+        "@dicetechnology/react-native-dice-video": "git+ssh://git@github.com:DiceTechnology/react-native-dice-video.git#6.43.1",
         "@dicetechnology/dice-unity": "^2.26.3",
         "keymirror": "0.1.1",
         "prop-types": "^15.5.10"
